Changed all if gain related items to type int

+++ wsa_lib.h
@@ -37,8 +37,8 @@ struct wsa_descriptor {
 	uint64_t max_tune_freq;
 	uint64_t min_tune_freq;
 	uint64_t freq_resolution;
-	float max_if_gain;
-	float min_if_gain;
+	int max_if_gain;
+	int min_if_gain;
 	float abs_max_amp[NUM_RF_GAINS];
 };
